// src/types/celebrity.ts
// src/types/celebrity.ts

// export interface Celebrity {
//   id: number;
//   first: string;
//   last: string;
//   dob: string;
//   gender: string;
//   email: string;
//   picture: string;
//   country: string;
//   description: string;
// }
// src/types/celebrity.ts
// types/celebrity.ts
export interface Celebrity {
  id: number;
  first: string;
  last: string;
  dob: string;
  gender: string;
  country: string;
  description: string;
}